Codesize is basically the same (a tiiiiny bit smaller):

```
Title                              Section             Old             New  Percent
WebURLBenchmark                     __text:        1418369         1418129    -0.0%
```

But I'm quite sure there is a non-fluke performance improvement for the common case when domain names do not include percent-encoding.

Before

```
name                                                   time         std        iterations warmup        
--------------------------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                 24860.000 ns ±  13.54 %      51832 3118901.000 ns
Constructor.SpecialNonFile.AverageURLs filtered        40453.000 ns ±  13.20 %      32265 4314659.000 ns
Constructor.SpecialNonFile.IPv4 host                   23958.000 ns ±  15.75 %      54709 2461829.000 ns
Constructor.SpecialNonFile.IPv4 host filtered          34855.000 ns ±  11.76 %      38040 3624207.000 ns
Constructor.SpecialNonFile.IPv6 host                   25037.000 ns ±  16.08 %      53182 2600285.000 ns
Constructor.SpecialNonFile.IPv6 host filtered          32421.000 ns ±  13.48 %      40545 3427260.000 ns
Constructor.SpecialNonFile.Percent-encoding components 10134.000 ns ±  16.55 %     119120 1148119.000 ns
Constructor.SpecialNonFile.Percent-encoded hostnames    8949.000 ns ±  14.63 %     149105  989243.000 ns
Constructor.SpecialNonFile.Long paths                  78507.000 ns ±   9.45 %      17017 8044446.000 ns
Constructor.SpecialNonFile.Complex paths 1             24138.000 ns ±   9.95 %      54900 2501131.000 ns
Constructor.SpecialNonFile.Complex paths 2             29436.000 ns ±  13.00 %      43328 3051703.000 ns
Constructor.SpecialNonFile.Long query 1                 2305.000 ns ±  27.23 %     545664  286796.000 ns
Constructor.SpecialNonFile.Long query 2                20301.000 ns ±  11.60 %      63377 2173905.000 ns
```

After

```
name                                                   time         std        iterations warmup         
---------------------------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                 22769.000 ns ±  25.03 %      58726  7695032.000 ns
Constructor.SpecialNonFile.AverageURLs filtered        43566.000 ns ±  26.18 %      28435 16173603.000 ns
Constructor.SpecialNonFile.IPv4 host                   22500.000 ns ±  25.00 %      54657  7267242.000 ns
Constructor.SpecialNonFile.IPv4 host filtered          36267.000 ns ±  25.12 %      33874 12638136.000 ns
Constructor.SpecialNonFile.IPv6 host                   26545.000 ns ±  26.07 %      47752  8714821.000 ns
Constructor.SpecialNonFile.IPv6 host filtered          34745.000 ns ±  25.12 %      35121 11060414.000 ns
Constructor.SpecialNonFile.Percent-encoding components 10065.000 ns ±  30.02 %     121421  4060102.000 ns
Constructor.SpecialNonFile.Percent-encoded hostnames    9544.000 ns ±  33.49 %     133956  2776224.000 ns
Constructor.SpecialNonFile.Long paths                  79031.000 ns ±  21.78 %      15783 28470589.000 ns
Constructor.SpecialNonFile.Complex paths 1             24579.000 ns ±  24.53 %      50223  8918264.000 ns
Constructor.SpecialNonFile.Complex paths 2             29694.000 ns ±  23.68 %      42279  9522208.000 ns
Constructor.SpecialNonFile.Long query 1                 2163.000 ns ±  46.99 %     560989   733054.000 ns
Constructor.SpecialNonFile.Long query 2                21286.000 ns ±  24.85 %      60591  6235494.000 ns
```

Also fixes an issue with the IDNA prefix matching, which should be case-insensitive.